chore(www): use filterDevRuntimeError for Monaco resize error
