+++ packages/brisa/src/utils/brisa-error-dialog/inject-code.ts
@@ -1,7 +1,6 @@
 import path from 'node:path';
 import clientBuildPlugin from '@/utils/client-build-plugin';
-import { logBuildError, logWarning } from '@/utils/log/log-build';
-import { getConstants } from '@/constants';
+import { logBuildError } from '@/utils/log/log-build';
 
 // Should be used via macro
 export async function injectBrisaDialogErrorCode() {
@@ -17,7 +16,7 @@ export async function injectBrisaDialogErrorCode() {
     target: 'browser',
     external: ['brisa'],
     define: {
-      __FILTER_DEV_RUNTIME_ERRORS__: getFilterDevRuntimeErrors(),
+      __FILTER_DEV_RUNTIME_ERRORS__: '__FILTER_DEV_RUNTIME_ERRORS__',
     },
     plugins: [
       {
@@ -42,22 +41,4 @@ export async function injectBrisaDialogErrorCode() {
   }
 
   return (await outputs?.[0]?.text?.()) ?? '';
-}
-
-export function getFilterDevRuntimeErrors() {
-  const { CONFIG } = getConstants();
-  const filterType = typeof CONFIG.filterRuntimeDevErrors;
-
-  if (filterType === 'function') {
-    return CONFIG.filterRuntimeDevErrors!.toString();
-  }
-
-  if (filterType !== 'undefined') {
-    logWarning(
-      ['CONFIG.filterRuntimeDevErrors should be a function'],
-      'Docs: https://brisa.build/building-your-application/configuring/filter-runtime-dev-errors',
-    );
-  }
-
-  return '() => true';
 }
