+++ packages/www/brisa.config.ts
@@ -4,4 +4,7 @@ import vercel from 'brisa-adapter-vercel';
 export default {
   output: 'static',
   outputAdapter: vercel(),
+  filterRuntimeDevErrors: (e) => {
+    return !e.message?.startsWith?.('ResizeObserver') ?? true;
+  },
 } satisfies Configuration;
